Temporary refactor to test on AWS
